% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{as.mcmc.metropolis.samples}
\alias{as.mcmc.metropolis.samples}
\title{Convert glm_metropolis output to `mcmc` object from package coda}
\usage{
\method{as.mcmc}{metropolis.samples}(object, ...)
}
\arguments{
\item{object}{an object from the function "metropolis"}

\item{...}{not used}
}
\description{
Allows use of useful functions from `coda` package
}
\details{
TBA
}
\examples{
runif(1)
}
