% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{expit}
\alias{expit}
\title{Inverse logit transform}
\usage{
expit(mu)
}
\arguments{
\item{mu}{log-odds}
}
\value{
returns a scalar or vector the same length 
as mu with values that are the inverse logit transform 
of mu
}
\description{
Inverse logit transform
}
\examples{
logodds = rnorm(10)
expit(logodds)
logodds = log(1.0)
expit(logodds)
}
