% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{magfields}
\alias{magfields}
\title{A case control study of childhood leukemia and magnetic fields}
\format{A data frame with 234 rows and 2 variables:
\describe{
  \item{y}{childhood leukemia}
  \item{x}{exposure to magnetic field}
}}
\usage{
magfields
}
\description{
A case control study of childhood leukemia and magnetic fields
}
\keyword{datasets}
