% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot.metropolis.samples}
\alias{plot.metropolis.samples}
\title{Plot the output from the metropolis function}
\usage{
\method{plot}{metropolis.samples}(x, keepburn = FALSE, parms = NULL,
  ...)
}
\arguments{
\item{x}{the outputted object from the "metropolis_glm" function}

\item{keepburn}{keep the burnin iterations in calculations (if adapt=TRUE, keepburn=TRUE}

\item{parms}{names of parameters to plot (plots the first by default, if TRUE, plots all)}

\item{...}{other arguments to plot}
}
\description{
This function allows you to summarize output from the metropolis function.
}
\details{
TBA
}
\examples{
runif(1)
}
