% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{print.metropolis.samples}
\alias{print.metropolis.samples}
\title{Print a metropolis.samples object}
\usage{
\method{print}{metropolis.samples}(x, ...)
}
\arguments{
\item{x}{a "metropolis.samples" object from the function "metropolis_glm"}

\item{...}{not used.}
}
\value{
An unmodified "metropolis.samples" object (invisibly)
}
\description{
This function allows you to summarize output from the "metropolis_glm" function.
}
\details{
None
}
