% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{summary.metropolis.samples}
\alias{summary.metropolis.samples}
\title{Summarize a probability distribution from a Markov Chain}
\usage{
\method{summary}{metropolis.samples}(object, keepburn = FALSE, ...)
}
\arguments{
\item{object}{an object from the function "metropolis"}

\item{keepburn}{keep the burnin iterations in calculations (if adapt=TRUE, keepburn=TRUE 
will yield potentially invalid summaries)}

\item{...}{not used}
}
\description{
This function allows you to summarize output from the metropolis function.
}
\details{
TBA
}
\examples{
runif(1)
}
