% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihoods.R
\name{logistic_ll}
\alias{logistic_ll}
\title{logistic log likelihood}
\usage{
logistic_ll(y, X, par)
}
\arguments{
\item{y}{binary outcome}

\item{X}{design matrix}

\item{par}{vector of model coefficients}
}
\value{
a scalar quantity proportional to a binomial likelihood
with logistic parameterization, given y,X,and par
}
\description{
logistic log likelihood
}
