% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihoods.R
\name{normal_ll}
\alias{normal_ll}
\title{Gaussian log likelihood}
\usage{
normal_ll(y, X, par)
}
\arguments{
\item{y}{binary outcome}

\item{X}{design matrix}

\item{par}{vector of gaussian scale parameter followed by model coefficients}
}
\value{
a scalar quantity proportional to a normal likelihood
with linear parameterization, given y, X, and par
}
\description{
Gaussian log likelihood
}
