% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot.metropolis.samples}
\alias{plot.metropolis.samples}
\title{Plot the output from the metropolis function}
\usage{
\method{plot}{metropolis.samples}(x, keepburn = FALSE, parms = NULL, ...)
}
\arguments{
\item{x}{the outputted object from the "metropolis_glm" function}

\item{keepburn}{keep the burnin iterations in calculations (if adapt=TRUE, keepburn=TRUE}

\item{parms}{names of parameters to plot (plots the first by default, if TRUE, plots all)}

\item{...}{other arguments to plot}
}
\value{
None
}
\description{
This function allows you to summarize output from the metropolis function.
}
\details{
TBA
}
\examples{
dat = data.frame(y = rbinom(100, 1, 0.5), x1=runif(100), x2 = runif(100))
res = metropolis_glm(y ~ x1 + x2, data=dat, family=binomial(), iter=10000, burnin=3000, 
adapt=TRUE, guided=TRUE, block=FALSE)
plot(res)
}
